+++ src/macros/mod.rs
@@ -1,0 +1,4 @@
+#[macro_use]
+pub mod serialize;
+#[macro_use]
+pub mod deserialize;