+++ src/prelude.rs
@@ -1,3 +1,3 @@
 pub use crate::hessian::{HDeserialize, HSerialize};
 pub use crate::serde::{Hessian, HessianWriteable};
-pub use hessian2_derive::HessianSerialize;
+pub use hessian2_derive::{HessianDeserialize, HessianSerialize};